+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -19,7 +19,7 @@ object Libs {
 
     private object Versions {
         const val junitJupiter = "5.8.2"
-        const val kotlinCoroutines = "1.5.2"
+        const val kotlinCoroutines = "1.6.0"
         const val moshi = "1.12.0"
         const val okhttp = "3.12.13"
         const val retrofit = "2.6.4"
@@ -29,7 +29,7 @@ object Libs {
 
     const val junitJupiterApi = "org.junit.jupiter:junit-jupiter-api:${Versions.junitJupiter}"
     const val junitJupiterEngine = "org.junit.jupiter:junit-jupiter-engine:${Versions.junitJupiter}"
-    const val kotlinCoroutinesCore = "org.jetbrains.kotlinx:kotlinx-coroutines-core:${Versions.kotlinCoroutines}"
+    const val kotlinCoroutinesTest = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.kotlinCoroutines}"
     const val moshi = "com.squareup.moshi:moshi:${Versions.moshi}"
     const val moshiKotlinCodegen = "com.squareup.moshi:moshi-kotlin-codegen:${Versions.moshi}"
     const val okhttp = "com.squareup.okhttp3:okhttp:${Versions.okhttp}"